requirements.txt
numpy
pywebview
django
